support windows and install specified version

+++ internal/install.go
@@ -10,7 +10,6 @@ import (
 	"log"
 	"os"
 	"os/exec"
-	"path/filepath"
 	"time"
 )
 
@@ -22,38 +21,24 @@ func Install(version string) error {
 	}
 	vinfos := versions[version]
 	if len(vinfos) == 0 {
-		return fmt.Errorf("version %q not found", version)
+		// 用于支持安装 3 位版本，如  go1.16.0、go1.16.3
+		err = installVV(version, versions)
+		if err == nil {
+			return nil
+		}
+		return fmt.Errorf("install %q failed: %w", version, err)
 	}
 	last := vinfos[0]
 
 	log.Println("[install]", "found last", version, "version is", last.Raw)
 
-	defer os.Chdir(TmpDir())
+	goBinTo := last.RawGoBinPath()
 
-	if err = os.Chdir(last.DlDir()); err != nil {
-		return err
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
-	defer cancel()
-
-	goBinTo := filepath.Join(GOBIN(), last.Raw)
-	cmd := exec.CommandContext(ctx, "go", "build", "-o", goBinTo)
-	log.Println("[exec]", cmd.String())
-	cmd.Stderr = os.Stderr
-	cmd.Stdout = os.Stdout
-	if err = cmd.Run(); err != nil {
+	if err = installWithVersion(last); err != nil {
 		return err
 	}
 
-	downloadCmd := exec.Command(goBinTo, "download")
-	log.Println("[exec]", downloadCmd.String())
-	downloadCmd.Stderr = os.Stderr
-	downloadCmd.Stdout = os.Stdout
-	if err = downloadCmd.Run(); err != nil {
-		return err
-	}
-
-	goBinLink := filepath.Join(GOBIN(), last.Normalized)
+	goBinLink := last.NormalizedGoBinPath()
 	if goBinLink == goBinTo {
 		return nil
 	}
@@ -64,17 +49,22 @@ func Install(version string) error {
 		if err = os.Remove(goBinLink); err != nil && !os.IsNotExist(err) {
 			return err
 		}
-
-		if err = os.Symlink(goBinTo, goBinLink); err != nil {
-			return err
+		if isWindows() {
+			if err = copyFile(goBinTo, goBinLink); err != nil {
+				return err
+			}
+		} else {
+			if err = os.Symlink(goBinTo, goBinLink); err != nil {
+				return err
+			}
 		}
 		log.Println("[link]", goBinTo, "->", goBinLink, "success")
 	}
 
 	// create sdk dir link
-	{
+	if !isWindows() {
 		sdkDir := last.Raw
-		sdkDirLink := mustGoRoot(last.Normalized)
+		sdkDirLink := mustGoRoot(last.Normalized) + ".latest"
 
 		if err = os.Remove(sdkDirLink); err != nil && !os.IsNotExist(err) {
 			return err
@@ -87,4 +77,56 @@ func Install(version string) error {
 	}
 	log.Printf("Success. You may now run '%s'\n", version)
 	return nil
+}
+
+func installWithVersion(ver *Version) error {
+	defer os.Chdir(TmpDir())
+	log.Println("[chdir]", ver.DlDir())
+	if err := os.Chdir(ver.DlDir()); err != nil {
+		return err
+	}
+	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
+	defer cancel()
+
+	goBinTo := ver.RawGoBinPath()
+	cmd := exec.CommandContext(ctx, "go", "build", "-o", goBinTo)
+	log.Println("[exec]", cmd.String())
+	cmd.Stderr = os.Stderr
+	cmd.Stdout = os.Stdout
+	if err := cmd.Run(); err != nil {
+		return err
+	}
+
+	downloadCmd := exec.Command(goBinTo, "download")
+	log.Println("[exec]", downloadCmd.String())
+	downloadCmd.Stderr = os.Stderr
+	downloadCmd.Stdout = os.Stdout
+	return downloadCmd.Run()
+}
+
+// installVV 安装指定的小版本
+func installVV(version string, vvs map[string][]*Version) error {
+	if len(vvs[version]) > 0 {
+		// 不应该执行到这个逻辑
+		return fmt.Errorf("now allow, bug here")
+	}
+	vu, err := parserVersion(version)
+	if err != nil {
+		return err
+	}
+	mvs := vvs[vu.Normalized]
+	if len(mvs) == 0 {
+		return fmt.Errorf("minor version not found")
+	}
+	var installVersion *Version
+	for _, mv := range mvs {
+		if mv.Raw == version || mv.Raw+".0" == version {
+			installVersion = mv
+			break
+		}
+	}
+	if installVersion == nil {
+		return fmt.Errorf("version not found")
+	}
+	return installWithVersion(installVersion)
 }
